+++ src/infrastructure/event/listeners/listeners.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { NotificationListener } from 'src/infrastructure/listeners/notification/notification.listener';
+import { NotificationListener } from 'src/infrastructure/event/listeners/notification/notification.listener';
 
 @Module({
   providers: [NotificationListener],
